changed the ReqPkt struct to hold 16bit values, TODO fix slave sensor code to format 2 byte values

+++ MpiCom/sharedi2cCom.h
@@ -15,11 +15,11 @@
 	 * as an extra parameter for the Pi to the Arduino
          */
 	typedef struct {
-                uint8_t header; //Command to the microcontroller
-                uint8_t throttle; 	//speed
-		uint8_t yaw;		//left & right
-		uint8_t pitch;		//forward and back
-		uint8_t roll;		//
+                uint16_t header; //Command to the microcontroller
+                uint16_t throttle; 	//speed
+		uint16_t yaw;		//left & right
+		uint16_t pitch;		//forward and back
+		uint16_t roll;		//
 	} ReqPkt;
 
 	/**
